Move wizard constants/helpers to platform
