+++ src/platform/site-wide/wizard/constants.js
@@ -1,0 +1,5 @@
+export const WIZARD_STATUS = 'wizardStatus';
+export const WIZARD_STATUS_NOT_STARTED = 'not started';
+export const WIZARD_STATUS_RESTARTING = 'restarting';
+export const WIZARD_STATUS_RESTARTED = 'restarted';
+export const WIZARD_STATUS_COMPLETE = 'complete';
