+++ src/platform/site-wide/wizard/index.js
@@ -1,0 +1,17 @@
+import { restartShouldRedirect } from './helpers';
+import {
+  WIZARD_STATUS,
+  WIZARD_STATUS_NOT_STARTED,
+  WIZARD_STATUS_RESTARTING,
+  WIZARD_STATUS_RESTARTED,
+  WIZARD_STATUS_COMPLETE,
+} from './constants';
+
+export default {
+  restartShouldRedirect,
+  WIZARD_STATUS,
+  WIZARD_STATUS_NOT_STARTED,
+  WIZARD_STATUS_RESTARTING,
+  WIZARD_STATUS_RESTARTED,
+  WIZARD_STATUS_COMPLETE,
+};
